lambda function 変更
